requirements.txt
dash==2.4.1
dash_bootstrap_components==1.1.0
dash_bootstrap_templates==1.0.5
dash_labs==1.0.8
numpy==1.22.3
pandas==1.4.3
plotly==5.8.0
python-dotenv==0.20.0
gunicorn==20.1.0
